Refactoring: common code moved from cc_history into the new module history_generator
